requirements.txt
asgiref==3.8.1
black==24.10.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Django
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
flake8==7.1.2
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
psutil==6.1.1
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
pytest==8.3.5
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
taskipy==1.14.1
tomli==2.2.1
uritemplate==4.1.1
urllib3==2.3.0
tensorflow
scikit-learn
h5py
pandas